Move UI components to line plot setting widget
